Fixed confusion in baseURL at request config section
```js
url: '/user',
baseURL: 'https://some-domain.com/api/'
                                     ^
```

this extra slash was creating confusion about the actual behavior of the config. about whether this Slash is necessary or not. In reality it is not necessary when used with the previous line `url: '/user'`. So, to avoid confusion one of the Slashes is needed to be removed. I removed the slash in `baseURL` cause it looks better in actuall usage of code.
Like the following line
```js
axios.get('/user/id');
```
Looks better than
```js
axios.get('user/id');
```