[Add] Kurdish language configuration file
Add `ku.lang.js` to add kurdish support for the project and also
introduced two new `t` properties which are `Next` and `Previous`
these are used in the `post.ejs` template to link current document
to it's previous and next pages, though it was hardcoded.
I added it in the Kurdish config file and added both keys in the
`en.lang.js` as it's the base file for new languages and current
langaues have to align themselves in it's regard.
Also, added two new configuration fields `dir` and `lang`
for `en.lang.js` as the base file, and `ku.lang.file`.
These two properties will be used in the `lang` and `dir`
html tag properties respectively.